introduce SubscanV2 API
