+++ subscrape/db/sqlitedict_wrapper.py
@@ -6,14 +6,12 @@ class SqliteDictWrapper:
     A wrapper for the sqlitedict library.
     """
 
-    def __init__(self, path, log_description, index_for_item):
+    def __init__(self, path, log_description):
         """
         :param path: str: the path to the sqlite database
-        :param index_for_item: function: a function that returns the index of an item
         :param log_description: str: a description of the specific database for logging purposes
         """
         self.logger = logging.getLogger(__name__)
-        self._index_for_item = index_for_item
         self.db = SqliteDict(path)
         self._write_count = 0
         self._log_description = log_description
@@ -24,8 +22,13 @@ class SqliteDictWrapper:
     def __exit__(self, exc_type, exc_val, exc_tb):
         self.db.close()
 
-    def write_item(self, data):
-        index = self._index_for_item(data)
+    def write_item(self, index, data) -> bool:
+        """
+        Write a single item to the database.
+        :param index: str: the index of the item to write
+        :param data: dict: the data to write
+        :return: bool: True if the item was new, False if it was already in the database
+        """
         was_new_element = index not in self.db
         self.db[index] = data
