+++ tests/test_events.py
@@ -1,9 +1,11 @@
 from subscrape.db.subscrape_db import SubscrapeDB
 import logging
 import subscrape
+import pytest
 
-def test():
-        
+@pytest.mark.parametrize("scraper", [(None), ("SubscanV2")])
+def test(scraper):
+    
     config = {
         "kusama":{
             "events":{
@@ -11,6 +13,9 @@ def test():
             }
         }
     }
+
+    if scraper is not None:
+        config["kusama"]["_scraper"] = scraper
 
     logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(name)s - %(message)s')
 
@@ -26,5 +31,3 @@ def test():
 
     first_crowdloan = events["14215808-27"]
     assert first_crowdloan["extrinsic_hash"] == '0x9d3430cd00bff235d4cdd595513375e6ceacb5228590a8629a865922d67f056f'
-
-test()
