+++ tests/test_extrinsics.py
@@ -1,8 +1,10 @@
 from subscrape.db.subscrape_db import SubscrapeDB
 import logging
 import subscrape
+import pytest
 
-def test():
+@pytest.mark.parametrize("scraper", [None, "SubscanV2"])
+def test(scraper):
         
     config = {
         "kusama":{
@@ -12,6 +14,9 @@ def test():
         }
     }
 
+    if scraper is not None:
+        config["kusama"]["_scraper"] = scraper
+    
     logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(name)s - %(message)s')
 
     logging.info("wiping storage")
@@ -27,4 +32,3 @@ def test():
     first_crowdloan = extrinsics["8974101-2"]
     assert first_crowdloan["extrinsic_hash"] == '0xee88a0694a88435c11b5f5d2ac971ab026180467ab215887b6608731c4679051'
 
-test()
