+++ tests/test_persistence.py
@@ -1,8 +1,10 @@
 from subscrape.db.subscrape_db import SubscrapeDB
 import logging
 import subscrape
+import pytest
 
-def test():
+@pytest.mark.parametrize("scraper", [None, "SubscanV2"])
+def test(scraper):
         
     config = {
         "kusama":{
@@ -15,6 +17,9 @@ def test():
         },
     }
 
+    if scraper is not None:
+        config["kusama"]["_scraper"] = scraper
+    
     logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(name)s - %(message)s')
 
     logging.info("wiping storage")
@@ -24,5 +29,3 @@ def test():
     items_scraped2 = subscrape.scrape(config)
     
     assert items_scraped1 != items_scraped2
-
-test()
